# Refactor campaigns go tests