--- a/enterprise/internal/campaigns/main_test.go
+++ b/enterprise/internal/campaigns/main_test.go
@@ -11,10 +11,7 @@ import (
 
 	"github.com/inconshreveable/log15"
 	"github.com/keegancsmith/sqlf"
-	"github.com/sourcegraph/go-diff/diff"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns/store"
-	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/campaigns"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
@@ -75,93 +72,6 @@ func truncateTables(t *testing.T, db *sql.DB, tables ...string) {
 	if err != nil {
 		t.Fatal(err)
 	}
-}
-
-type testSpecOpts struct {
-	user         int32
-	repo         api.RepoID
-	campaignSpec int64
-
-	// If this is non-blank, the changesetSpec will be an import/track spec for
-	// the changeset with the given externalID in the given repo.
-	externalID string
-
-	// If this is set, the changesetSpec will be a "create commit on this
-	// branch" changeset spec.
-	headRef string
-
-	// If this is set along with headRef, the changesetSpec will have published
-	// set.
-	published interface{}
-
-	title             string
-	body              string
-	commitMessage     string
-	commitDiff        string
-	commitAuthorEmail string
-	commitAuthorName  string
-}
-
-var testChangsetSpecDiffStat = &diff.Stat{Added: 10, Changed: 5, Deleted: 2}
-
-func buildChangesetSpec(t *testing.T, opts testSpecOpts) *campaigns.ChangesetSpec {
-	t.Helper()
-
-	published := campaigns.PublishedValue{Val: opts.published}
-	if opts.published == nil {
-		// Set false as the default.
-		published.Val = false
-	}
-	if !published.Valid() {
-		t.Fatalf("invalid value for published passed, got %v (%T)", opts.published, opts.published)
-	}
-
-	spec := &campaigns.ChangesetSpec{
-		UserID:         opts.user,
-		RepoID:         opts.repo,
-		CampaignSpecID: opts.campaignSpec,
-		Spec: &campaigns.ChangesetSpecDescription{
-			BaseRepository: graphqlbackend.MarshalRepositoryID(opts.repo),
-
-			ExternalID: opts.externalID,
-			HeadRef:    opts.headRef,
-			Published:  published,
-
-			Title: opts.title,
-			Body:  opts.body,
-
-			Commits: []campaigns.GitCommitDescription{
-				{
-					Message:     opts.commitMessage,
-					Diff:        opts.commitDiff,
-					AuthorEmail: opts.commitAuthorEmail,
-					AuthorName:  opts.commitAuthorName,
-				},
-			},
-		},
-		DiffStatAdded:   testChangsetSpecDiffStat.Added,
-		DiffStatChanged: testChangsetSpecDiffStat.Changed,
-		DiffStatDeleted: testChangsetSpecDiffStat.Deleted,
-	}
-
-	return spec
-}
-
-func createChangesetSpec(
-	t *testing.T,
-	ctx context.Context,
-	store *store.Store,
-	opts testSpecOpts,
-) *campaigns.ChangesetSpec {
-	t.Helper()
-
-	spec := buildChangesetSpec(t, opts)
-
-	if err := store.CreateChangesetSpec(ctx, spec); err != nil {
-		t.Fatal(err)
-	}
-
-	return spec
 }
 
 func createCampaignSpec(t *testing.T, ctx context.Context, store *store.Store, name string, userID int32) *campaigns.CampaignSpec {