--- a/enterprise/internal/campaigns/main_test.go
+++ b/enterprise/internal/campaigns/main_test.go
@@ -4,18 +4,14 @@ import (
 	"context"
 	"database/sql"
 	"flag"
-	"fmt"
 	"os"
 	"strings"
 	"testing"
 
 	"github.com/inconshreveable/log15"
-	"github.com/keegancsmith/sqlf"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns/store"
 	"github.com/sourcegraph/sourcegraph/internal/campaigns"
-	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
-	"github.com/sourcegraph/sourcegraph/internal/types"
 )
 
 func init() {
@@ -29,41 +25,6 @@ func TestMain(m *testing.M) {
 	}
 	os.Exit(m.Run())
 }
-
-var createTestUser = func() func(*testing.T, bool) *types.User {
-	count := 0
-
-	// This function replicates the minium amount of work required by
-	// db.Users.Create to create a new user, but it doesn't require passing in
-	// a full db.NewUser every time.
-	return func(t *testing.T, siteAdmin bool) *types.User {
-		t.Helper()
-
-		user := &types.User{
-			Username:    fmt.Sprintf("testuser-%d", count),
-			DisplayName: "testuser",
-		}
-
-		q := sqlf.Sprintf("INSERT INTO users (username, site_admin) VALUES (%s, %t) RETURNING id, site_admin", user.Username, siteAdmin)
-		err := dbconn.Global.QueryRow(q.Query(sqlf.PostgresBindVar), q.Args()...).Scan(&user.ID, &user.SiteAdmin)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if user.SiteAdmin != siteAdmin {
-			t.Fatalf("user.SiteAdmin=%t, but expected is %t", user.SiteAdmin, siteAdmin)
-		}
-
-		_, err = dbconn.Global.Exec("INSERT INTO names(name, user_id) VALUES($1, $2)", user.Username, user.ID)
-		if err != nil {
-			t.Fatalf("failed to create name: %s", err)
-		}
-
-		count += 1
-
-		return user
-	}
-}()
 
 func truncateTables(t *testing.T, db *sql.DB, tables ...string) {
 	t.Helper()