--- a/enterprise/internal/campaigns/resolvers/campaign_connection_test.go
+++ b/enterprise/internal/campaigns/resolvers/campaign_connection_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/campaigns"
 	"github.com/sourcegraph/sourcegraph/internal/db"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
+	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/repos"
 )
 
@@ -27,7 +28,7 @@ func TestCampaignConnectionResolver(t *testing.T) {
 	}
 
 	ctx := backend.WithAuthzBypass(context.Background())
-	setupGlobalTestDB(t)
+	dbtesting.SetupGlobalTestDBWithoutReset(t)
 
 	userID := ct.CreateTestUser(t, true).ID
 
@@ -179,7 +180,7 @@ func TestCampaignsListing(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	setupGlobalTestDB(t)
+	dbtesting.SetupGlobalTestDBWithoutReset(t)
 
 	userID := ct.CreateTestUser(t, true).ID
 	actorCtx := actor.WithActor(ctx, actor.FromUser(userID))