@eseliger 2020-12-14
I know why it breaks - because some names are hard-coded and cause naming collisions. I will try to address those, then it should work. No strange global state though :)